+++ src/templates/add_topic_page.html
@@ -16,7 +16,7 @@
           </div>
           <!-- Column for the search form, aligning it to the right -->
           <div class="col-auto">
-            <form action="/search_forum" method="get" class="d-flex">
+            <form action="/forum/search_forum" method="get" class="d-flex">
               <input type="text" name="query" placeholder="Search topics ..." required class="form-control me-2">
               <input type="submit" value="Search" class="btn" style="background-color: #a53045; color: white; border: none;">
             </form>
