+++ src/templates/feedback_page.html
@@ -69,7 +69,7 @@ Feedback Page
         <div class="feedback-form-container">
             <button style="cursor: pointer;" id="createToView">← View Submitted Feedback</button>
             <h2 style="color: #941035; margin-bottom: 15px;">Feedback & Suggestions</h2>
-            <form id="feedbackForm" action="{{ url_for('submit_feedback') }}" method="POST">
+            <form id="feedbackForm" action="{{ url_for('feedback.submit_feedback') }}" method="POST">
                 <label for="name" style="color: #941035; margin-bottom: 15px;">Your Name (optional):</label>
                 <input type="text" id="name" name="name">
         
@@ -224,7 +224,7 @@ Feedback Page
             e.preventDefault();
             const formData = new FormData(this);
             
-            fetch("{{ url_for('submit_feedback') }}", {
+            fetch("{{ url_for('feedback.submit_feedback') }}", {
                 method: 'POST',
                 body: formData
             })
