+++ src/templates/forum_page.html
@@ -55,7 +55,7 @@
                   Posted by <span class="topic-author">@{{username}}</span>
                 </div>
                 <h2 class="topic-title">
-                  <a style="color: #941035;" href="/forum/topic/{{ topic.id }}">{{ topic.title }}</a>
+                  <a style="color: #941035;" href="/fm/topic/{{ topic.id }}">{{ topic.title }}</a>
                 </h2><div style="color: #555555;"bclass="topic-date">Posted on: {{ topic.date }}</div>
                 <div class="topic-tag">{{ topic.tag }}</div>
                 {% if topic.imageUrl and topic.imageUrl != 'none' %}
