+++ src/templates/header.html
@@ -11,11 +11,11 @@
                                 class="glyphicon glyphicon-home" title="home page"
                                 style="font-size:24px; margin-left: 30px;"></i></a>
                 <a class="tasks {{ 'active' if current_page == 'tasks' else '' }}"
-                        href="{{ url_for('tasks_page') }}"><i class="glyphicon glyphicon-list-alt" title="tasks"
+                        href="{{ url_for('tasks.tasks_page') }}"><i class="glyphicon glyphicon-list-alt" title="tasks"
                                 style="font-size:24px;"></i>
                 </a>
                 <a class="timer {{ 'active' if current_page == 'pomodoro_page' else '' }}"
-                        href="{{ url_for('pomodoro_page') }}"><i class="glyphicon glyphicon-time" title="pomodoro timer"
+                        href="{{ url_for('pomodoro.pomodoro_page') }}"><i class="glyphicon glyphicon-time" title="pomodoro timer"
                                 style="font-size:24px;"></i>
                 </a>
                 <a class="courses {{ 'active' if current_page == 'course_page' else '' }}"
@@ -23,7 +23,7 @@
                                 style="font-size:24px;"></i>
                 </a>
                 <a class="feedback {{ 'active' if current_page == 'feedback_page' else '' }}"
-                        href="{{ url_for('feedback_page') }}"><i class="glyphicon glyphicon-comment"
+                        href="{{ url_for('feedback.feedback_page') }}"><i class="glyphicon glyphicon-comment"
                                 title="feedback box" style="font-size:24px;"></i>
                 </a>
                 <a class="forum {{ 'active' if current_page == 'forum_page' else '' }}"
@@ -50,14 +50,14 @@
             <li class="nav-item {{ 'active' if request.path == url_for('courses.course_page') else '' }}">
                 <a class="nav-link" href="{{ url_for('courses.course_page') }}">Courses</a>
               </li>
-            <li class="nav-item {{ 'active' if request.path == url_for('tasks_page') else '' }}">
-                <a class="nav-link" href="{{ url_for('tasks_page') }}"> Tasks</a>
+            <li class="nav-item {{ 'active' if request.path == url_for('tasks.tasks_page') else '' }}">
+                <a class="nav-link" href="{{ url_for('tasks.tasks_page') }}"> Tasks</a>
               </li>
-            <li class="nav-item {{ 'active' if request.path == url_for('pomodoro_page') else '' }}">
-              <a class="nav-link" href="{{ url_for('pomodoro_page') }}">Pomodoro Timer</a>
+            <li class="nav-item {{ 'active' if request.path == url_for('pomodoro.pomodoro_page') else '' }}">
+              <a class="nav-link" href="{{ url_for('pomodoro.pomodoro_page') }}">Pomodoro Timer</a>
             </li>
-            <li class="nav-item {{ 'active' if request.path == url_for('feedback_page') else '' }}">
-              <a class="nav-link" href="{{ url_for('feedback_page') }}">Feedback</a>
+            <li class="nav-item {{ 'active' if request.path == url_for('feedback.feedback_page') else '' }}">
+              <a class="nav-link" href="{{ url_for('feedback.feedback_page') }}">Feedback</a>
             </li>
             <li class="nav-item {{ 'active' if request.path == url_for('forum.forum_page') else '' }}">
               <a class="nav-link" href="{{ url_for('forum.forum_page') }}">Forum</a>
