+++ src/templates/tasks.html
@@ -54,7 +54,7 @@ href="{{ url_for('tasks') }}"><i class="glyphicon glyphicon-task" title="tasks"
                         </div>
                         <div class="priority-row">
                             <div class="task-priority {{ 'priority-' + task.priority|lower }}">{{ task.priority|capitalize }} Priority</div>
-                            <a href="{{ url_for('pomodoro_page', task_id=task.id, status='todo') }}">Start Pomodoro</a>
+                            <a href="{{ url_for('pomodoro.pomodoro_page', task_id=task.id, status='todo') }}">Start Pomodoro</a>
                         </div>
                     </div>
                     {% endfor %}
@@ -94,7 +94,7 @@ href="{{ url_for('tasks') }}"><i class="glyphicon glyphicon-task" title="tasks"
                 </div>
                 <div class="priority-row">
                     <div class="task-priority {{ 'priority-' + task.priority|lower }}">{{ task.priority|capitalize }} Priority</div>
-                    <a href="{{ url_for('pomodoro_page', task_id=task.id, status='in_progress') }}">Start Pomodoro</a>
+                    <a href="{{ url_for('pomodoro.pomodoro_page', task_id=task.id, status='in_progress') }}">Start Pomodoro</a>
                 </div>
             </div>
             {% endfor %}
