+++ src/util.py
@@ -9,7 +9,7 @@ import json
 import csv
 import openai
 from io import StringIO
-from datetime import datetime, timedelta, timezone
+from datetime import datetime
 from joblib import load
 from sklearn.base import TransformerMixin
 from sklearn.pipeline import Pipeline
